spiders: rename last_run_store module to lastrunstore_spider
