+++ tests/unit/test_oaipmh.py
@@ -12,7 +12,7 @@ from mock import patch
 import pytest
 
 from hepcrawl.spiders.common.oaipmh_spider import OAIPMHSpider
-from hepcrawl.spiders.common.last_run_store import NoLastRunToLoad
+from hepcrawl.spiders.common.lastrunstore_spider import NoLastRunToLoad
 from hepcrawl.testlib.fixtures import clean_dir
 from scrapy.utils.project import get_project_settings
